ml_service/requirements.txt
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
joblib==1.3.2
nltk==3.8.1 